requirements.txt
absl-py==1.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrdict==2.0.1
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
boto3==1.24.21
botocore==1.27.21
brotlipy==0.7.0
cachetools==5.2.0
catalogue==1.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cryptography==37.0.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
easydict==1.9
en-vectors-web-lg==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
fvcore==0.1.5.post20220512
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
iopath==0.1.9
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonlines==3.1.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-packaging==0.12.2
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
lmdb==1.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.5.3
murmurhash==1.0.7
nbclassic==0.4.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.2.2
open3d==0.15.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
plac==1.1.3
plyfile==0.7.4
pointnet2==3.0.0
pointops==0.0.0
portalocker==2.4.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.4
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
sentencepiece==0.1.96
setuptools==61.2.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==2.3.7
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==1.0.5
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.15.0
thinc==7.4.5
threadpoolctl==3.1.0
timm==0.6.5
tinycss2==1.1.1
tokenizers==0.12.1
tomlkit==0.11.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
trimesh==2.35.39
typer==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
warmup-scheduler==0.3
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.1
yacs==0.1.8
yapf==0.32.0
zipp==3.8.0
